# docs/openapi.yaml
openapi: 3.0.0
info:
  title: Encuesta Oficial Charros - Submit API
  version: 1.2.0
  description: API HTTP para recibir y exportar respuestas de la landing Encuesta Oficial Charros 2026-2027.

servers:
  - url: https://tu-api.onrender.com

paths:
  /healthz:
    get:
      summary: Estado de la API
      responses:
        '200':
          description: API disponible
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
                  status:
                    type: string
                  queueSize:
                    type: integer

  /api/submit:
    post:
      summary: Recibe una respuesta de la encuesta
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                timestamp:
                  type: string
                  format: date-time
                submissionId:
                  type: string
                campaignName:
                  type: string
                source:
                  type: string
                nombre:
                  type: string
                apellido:
                  type: string
                email:
                  type: string
                  format: email
                telefono:
                  type: string
                myCashlessId:
                  type: string
                rangoEdad:
                  type: string
                sexo:
                  type: string
                municipio:
                  type: string
                relacionCharros:
                  type: string
                antiguedad:
                  type: string
                acompanantes:
                  type: string
                motivacion:
                  type: string
                calificacionExperiencia:
                  type: string
                aspectosDisfrutados:
                  type: array
                  items:
                    type: string
                aspectosDisfrutadosOtro:
                  type: string
                aspectosMejorar:
                  type: string
                comentarioExperiencia:
                  type: string
                facilidadMyCashless:
                  type: string
                comentarioMyCashless:
                  type: string
                consumoEstadio:
                  type: string
                interesClubCharros:
                  type: string
                razonNoRenovo:
                  type: string
                beneficioPreferido:
                  type: array
                  items:
                    type: string
                canalPromociones:
                  type: array
                  items:
                    type: string
                tipoInformacion:
                  type: array
                  items:
                    type: string
                comentario:
                  type: string
                aceptaAvisoPrivacidad:
                  type: boolean
                aceptaComunicaciones:
                  type: boolean
              required:
                - nombre
                - apellido
                - email
      responses:
        '202':
          description: Respuesta encolada para persistencia
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
                  stored:
                    type: boolean
                  queued:
                    type: boolean
        '400':
          description: Faltan campos requeridos
        '503':
          description: Cola saturada, intenta de nuevo

  /api/submissions.csv:
    get:
      summary: Descarga de respuestas en CSV
      responses:
        '200':
          description: Archivo CSV con respuestas
          content:
            text/csv:
              schema:
                type: string

  /api/lead-submit:
    post:
      summary: Registra un lead de estadio (solo 1 por correo al día)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                timestamp:
                  type: string
                  format: date-time
                submissionId:
                  type: string
                campaignName:
                  type: string
                source:
                  type: string
                nombre:
                  type: string
                apellido:
                  type: string
                email:
                  type: string
                  format: email
                telefono:
                  type: string
                municipio:
                  type: string
                frecuenciaVisita:
                  type: string
                aceptaAvisoPrivacidad:
                  type: boolean
                aceptaComunicaciones:
                  type: boolean
                aceptaRegistroDiario:
                  type: boolean
              required:
                - nombre
                - apellido
                - email
                - telefono
                - municipio
                - frecuenciaVisita
                - aceptaAvisoPrivacidad
                - aceptaRegistroDiario
      responses:
        '202':
          description: Lead encolado para persistencia
        '400':
          description: Faltan campos requeridos
        '409':
          description: El correo ya registró un lead hoy
        '503':
          description: Cola saturada, intenta de nuevo

  /api/leads-submissions.csv:
    get:
      summary: Descarga de leads en CSV
      responses:
        '200':
          description: Archivo CSV con leads de estadio
          content:
            text/csv:
              schema:
                type: string